[safety-rules] Add a benchmark for process
Moved the generic process wrapper code into a test util so it can be
leveraged across the test and benchmark. The key point being that the
benchmark only wants a client, but we need to retain a handle on the
process or it will perish.

Also note that historically the benchmarks were /not/ correctly
executing.
